refactor: Separate draw module parts into folders
