+++ projects/hslayers/src/components/draw/draw-layer-metadata/draw-layer-metadata.html
