+++ projects/hslayers/src/components/draw/draw-toolbar/draw-toolbar.html
