+++ projects/hslayers/src/components/draw/draw.module.ts
@@ -7,8 +7,8 @@ import {TranslateModule} from '@ngx-translate/core';
 
 import {FormsModule} from '@angular/forms';
 import {HsDrawComponent} from './draw.component';
-import {HsDrawLayerMetadataDialogComponent} from './draw-layer-metadata.component';
-import {HsDrawToolbarComponent} from './draw-toolbar.component';
+import {HsDrawLayerMetadataDialogComponent} from './draw-layer-metadata/draw-layer-metadata.component';
+import {HsDrawToolbarComponent} from './draw-toolbar/draw-toolbar.component';
 import {HsQueryModule} from '../query/query.module';
 import {NgbDropdownModule} from '@ng-bootstrap/ng-bootstrap';
