+++ projects/hslayers/src/components/draw/public-api.ts
@@ -1,5 +1,5 @@
-export * from './draw-layer-metadata.component';
-export * from './draw-toolbar.component';
+export * from './draw-layer-metadata/draw-layer-metadata.component';
+export * from './draw-toolbar/draw-toolbar.component';
 export * from './draw.component';
 export * from './draw.module';
 export * from './draw.service';
