+++ projects/hslayers/src/hslayers.component.ts
@@ -4,7 +4,8 @@ import {HsAddDataComponent} from './components/add-data/add-data.component';
 import {HsCompositionsComponent} from './components/compositions/compositions.component';
 import {HsConfig} from './config.service';
 import {HsDrawComponent} from './components/draw/draw.component';
-import {HsDrawToolbarComponent} from './components/draw/draw-toolbar.component';
+import {HsDrawToolbarComponent} from './components/draw/draw-toolbar/draw-toolbar.component';
+import {HsFeatureInfoComponent} from './components/query/query-popup-feature/feature-widgets/feature-info.component';
 import {HsFeatureTableComponent} from './components/feature-table/feature-table.component';
 import {HsGeolocationComponent} from './components/geolocation/geolocation.component';
 import {HsInfoComponent} from './components/info/info.component';
